(travisci): fix travis build test errors
